Reworked the command-line argument parser to account for the new stricter flag parsing library.
Now, we first split the arguments slice into the slice before the word 'run' and the slice after the word 'run',
and we parse only the first slice.

Before, we would parse all of the arguments and have the flagging library stop when it encountered the word 'run', returning
the remaining flags as "arguments". Our new flagging library doesn't allow us to do this.